refactor(form-fields): split widget appearance parsing
Separate object-stream extraction and PDF string token decoding from widget caption scanning so the best-effort appearance parser stays easier to review without changing form-field output.
